+++ front/components/actions/process/ProcessActionDetails.tsx
@@ -11,8 +11,8 @@ import { useMemo } from "react";
 
 import { ActionDetailsWrapper } from "@app/components/actions/ActionDetailsWrapper";
 import type { ActionDetailsComponentBaseProps } from "@app/components/actions/types";
-import type { ProcessActionType } from "@app/lib/actions/types/process";
-import { PROCESS_ACTION_TOP_K } from "@app/lib/actions/types/process";
+import type { ProcessActionType } from "@app/lib/actions/process";
+import { PROCESS_ACTION_TOP_K } from "@app/lib/actions/process";
 
 export function ProcessActionDetails({
   action,
