+++ front/components/assistant_builder/actions/ProcessAction.tsx
@@ -24,7 +24,7 @@ import type {
   AssistantBuilderProcessConfiguration,
 } from "@app/components/assistant_builder/types";
 import { EmptyCallToAction } from "@app/components/EmptyCallToAction";
-import type { ProcessSchemaPropertyType } from "@app/lib/actions/types/process";
+import type { ProcessSchemaPropertyType } from "@app/lib/actions/process";
 import { classNames } from "@app/lib/utils";
 import type { Result, SpaceType, WorkspaceType } from "@app/types";
 import { Err, Ok } from "@app/types";
