+++ front/components/assistant_builder/server_side_props_helpers.ts
@@ -10,6 +10,7 @@ import {
 } from "@app/components/assistant_builder/types";
 import { REASONING_MODEL_CONFIGS } from "@app/components/providers/types";
 import type { DustAppRunConfigurationType } from "@app/lib/actions/dust_app_run";
+import type { ProcessConfigurationType } from "@app/lib/actions/process";
 import type { TablesQueryConfigurationType } from "@app/lib/actions/tables_query";
 import type { AgentActionConfigurationType } from "@app/lib/actions/types/agent";
 import {
@@ -21,7 +22,6 @@ import {
   isTablesQueryConfiguration,
   isWebsearchConfiguration,
 } from "@app/lib/actions/types/guards";
-import type { ProcessConfigurationType } from "@app/lib/actions/types/process";
 import type { ReasoningConfigurationType } from "@app/lib/actions/types/reasoning";
 import type { RetrievalConfigurationType } from "@app/lib/actions/types/retrieval";
 import { getContentNodesForDataSourceView } from "@app/lib/api/data_source_view";
