+++ front/lib/actions/runners.ts
@@ -3,6 +3,7 @@ import { BrowseConfigurationServerRunner } from "@app/lib/actions/browse";
 import { ConversationIncludeFileConfigurationServerRunner } from "@app/lib/actions/conversation/include_file";
 import type { DustAppRunConfigurationType } from "@app/lib/actions/dust_app_run";
 import { DustAppRunConfigurationServerRunner } from "@app/lib/actions/dust_app_run";
+import type { ProcessConfigurationType } from "@app/lib/actions/process";
 import { ProcessConfigurationServerRunner } from "@app/lib/actions/process";
 import { ReasoningConfigurationServerRunner } from "@app/lib/actions/reasoning";
 import { RetrievalConfigurationServerRunner } from "@app/lib/actions/retrieval";
@@ -16,7 +17,6 @@ import type {
   BaseActionConfigurationStaticMethods,
 } from "@app/lib/actions/types/base";
 import type { ConversationIncludeFileConfigurationType } from "@app/lib/actions/types/conversation/include_file";
-import type { ProcessConfigurationType } from "@app/lib/actions/types/process";
 import type { ReasoningConfigurationType } from "@app/lib/actions/types/reasoning";
 import type { RetrievalConfigurationType } from "@app/lib/actions/types/retrieval";
 import type { SearchLabelsConfigurationType } from "@app/lib/actions/types/search_labels";
